+++ CHANGELOG.md
@@ -7,6 +7,11 @@ This project <em>does not yet</em> adheres to [Semantic Versioning](https://semv
 
 ## [unreleased]
 
+### Added
+- Refresh event for smart selectors [#2255]
+
+[#2255]:https://github.com/SpeciesFileGroup/taxonworks/issues/2255
+
 \-
 
 ## [0.18.0] - 2020-05-14